successfully push to beacon from conductor
+ major refactor -- pull functions into 'util' files to allow for easier cross-package import
+ successfully push from beacon to conductor
