+++ purchasing/opportunities/util.py
@@ -1,9 +1,24 @@
 # -*- coding: utf-8 -*-
 
+import os
+import json
+
 from collections import defaultdict
 
-from purchasing.opportunities.forms import ValidationError
-from purchasing.opportunities.models import Category
+from flask import current_app
+
+from flask_login import current_user
+from werkzeug import secure_filename
+
+from purchasing.utils import (
+    connect_to_s3, _get_aggressive_cache_headers, random_id
+)
+from purchasing.database import db
+from purchasing.opportunities.forms import OpportunityForm
+from purchasing.opportunities.models import (
+    Opportunity, RequiredBidDocument, Category, OpportunityDocument
+)
+from purchasing.users.models import User, Role
 
 def get_categories(all_categories, form):
     '''Build category/subcategory lists/dictionaries
@@ -61,3 +76,128 @@ def fix_form_categories(request, form, cls, validate=None, obj=None,):
             form.errors[validate] = ['You must choose at least one!']
 
     return form_data
+
+def upload_document(document, _id):
+    if document is None or document == '':
+        return None, None
+
+    filename = secure_filename(document.filename)
+
+    if filename == '':
+        return None, None
+
+    _filename = 'opportunity-{}-{}'.format(_id, filename)
+
+    if current_app.config.get('UPLOAD_S3') is True:
+        # upload file to s3
+        conn, bucket = connect_to_s3(
+            current_app.config['AWS_ACCESS_KEY_ID'],
+            current_app.config['AWS_SECRET_ACCESS_KEY'],
+            current_app.config['UPLOAD_DESTINATION']
+        )
+        _document = bucket.new_key(_filename)
+        aggressive_headers = _get_aggressive_cache_headers(_document)
+        _document.set_contents_from_file(document, headers=aggressive_headers, replace=True)
+        _document.set_acl('public-read')
+        return _document.name, _document.generate_url(expires_in=0, query_auth=False)
+
+    else:
+        try:
+            os.mkdir(current_app.config['UPLOAD_DESTINATION'])
+        except:
+            pass
+
+        filepath = os.path.join(current_app.config['UPLOAD_DESTINATION'], _filename)
+        document.save(filepath)
+        return filepath
+
+def build_opportunity(data, publish=None, opportunity=None):
+    '''Create/edit a new opportunity
+
+    data - the form data from the request
+    publish - either 'publish' or 'save':
+      determines if the opportunity should be made public
+    opportunity - the actual opportunity object
+    '''
+    contact_email = data.pop('contact_email')
+    contact = User.query.filter(User.email == contact_email).first()
+
+    # pop off our documents so they don't
+    # get passed to the Opportunity constructor
+    documents = data.pop('documents')
+
+    if contact is None:
+        contact = User().create(
+            email=contact_email,
+            role=Role.query.filter(Role.name == 'staff').first(),
+            department=data.get('department')
+        )
+
+    _id = opportunity.id if opportunity else None
+
+    data.update(dict(contact_id=contact.id))
+
+    if opportunity:
+        opportunity = opportunity.update(**data)
+    else:
+        data.update(dict(created_by_id=current_user.id))
+        opportunity = Opportunity.create(**data)
+
+    opp_documents = opportunity.opportunity_documents.all()
+
+    for document in documents.entries:
+        if document.title.data == '':
+            continue
+
+        _id = _id if _id else random_id(6)
+
+        _file = document.document.data
+        if _file.filename in [i.name for i in opp_documents]:
+            continue
+
+        filepath = upload_document(_file, _id)
+        if filepath:
+            opportunity.opportunity_documents.append(OpportunityDocument(
+                name=document.title.data, href=filepath
+            ))
+
+    if not opportunity.is_public:
+        opportunity.is_public = True if publish == 'publish' else False
+
+    db.session.commit()
+    return opportunity
+
+def generate_opportunity_form(obj=None, form=OpportunityForm):
+    all_categories = Category.query.all()
+    form = form(obj=obj)
+
+    categories, subcategories, form = get_categories(all_categories, form)
+    display_categories = subcategories.keys()
+    if 'Select All' in display_categories:
+        display_categories.remove('Select All')
+
+    form.vendor_documents_needed.choices = [i.get_choices() for i in RequiredBidDocument.query.all()]
+
+    return form, json.dumps(sorted(display_categories) + ['Select All']), json.dumps(subcategories)
+
+def build_downloadable_groups(val, iterable):
+    '''Sorts and dedupes related lists
+
+    Handles quoting, deduping, and sorting
+    '''
+    return '"' + '; '.join(
+        sorted(list(set([i.__dict__[val] for i in iterable])))
+    ) + '"'
+
+def build_vendor_row(vendor):
+    '''Takes in a vendor and returns a list of that vendor's properties
+
+    Used to build the signup csv download
+    '''
+    return [
+        vendor.first_name, vendor.last_name, vendor.business_name,
+        vendor.email, vendor.phone_number, vendor.minority_owned,
+        vendor.woman_owned, vendor.veteran_owned, vendor.disadvantaged_owned,
+        build_downloadable_groups('category_friendly_name', vendor.categories),
+        build_downloadable_groups('title', vendor.opportunities)
+    ]
