+++ purchasing/static/js/shared/categories.js
