+++ purchasing/static/less/conductor/_detail_view.less
@@ -10,6 +10,18 @@
   margin-top: 15px;
 }
 
+.event-well {
+  position: fixed;
+}
+
 .js-email-contact-card {
   cursor: pointer;
+}
+
+.checkbox input[type="checkbox"] {
+  margin-left: 0;
+}
+
+.checkbox label {
+  padding-left: 25px;
 }
