+++ purchasing/static/less/opportunities/_styles.less
@@ -50,6 +50,14 @@ h4 {
   padding-top: 7px;
 }
 
+.checkbox input[type="checkbox"] {
+  margin-left: 0;
+}
+
+.checkbox label {
+  padding-left: 25px;
+}
+
 .navbar-text {
   @media @mobile {
     margin-left: 15px
@@ -148,14 +156,6 @@ h4 {
   p {
     line-height: 1.5;
   }
-}
-
-.checkbox input[type="checkbox"] {
-  margin-left: 0;
-}
-
-.checkbox label {
-  padding-left: 25px;
 }
 
 .another-category {
