+++ purchasing/templates/conductor/detail.html
@@ -11,7 +11,7 @@
       {% include "conductor/detail/event_well.html" %}
    </div><!-- event timeline view -->
 
-    <div class="col-sm-8">
+    <div class="col-sm-7 col-sm-offset-5">
       {% include 'conductor/detail/title_card.html' %}
 
       <h4>Activity Stream</h4>
